feat: endorse claims against other identities
Until now only the user with the private key was able to create claims
and endorsements against the identity.
Introduce the concept to PrivateIdentity and PublicIdentity. A
PrivateIdentity is used to authenticate against the iov42 platform,
authorise requests, and endorse claims against other entities.
The PublicIdentity is used to interact with the platform to publish the
identity identifier and the public key or to create claims/endorsements
against it. This way 3rd parties can use the PublicIdentity to create
endorsements against the identity.